moving audio conversion to a separate pipeline
