+++ ChildProject/pipelines/__init__.py
@@ -1,1 +1,2 @@
 from .zooniverse import ZooniversePipeline
+from .conversion import ConversionPipeline
