+++ ChildProject/pipelines/pipeline.py
@@ -4,11 +4,9 @@ class Pipeline(ABC):
     def __init__(self):
         pass
 
-    @abstractmethod
     def check_setup(self):
         pass
 
-    @abstractmethod
     def setup(self):
         pass
